backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-CORS==4.0.0
Werkzeug==2.3.6
requests==2.31.0
python-dotenv==1.0.0
openai==0.27.8
razorpay==1.3.0
